NoSQL Spike

+++ packages/syft/src/syft/core/tensor/tensor.py
@@ -169,7 +169,7 @@ class TensorPointer(Pointer):
             other_dtype = other.public_dtype
         elif isinstance(other, (int, float)):
             other_shape = (1,)
-            other_dtype = np.int32
+            other_dtype = np.int64
         elif isinstance(other, bool):
             other_shape = (1,)
             other_dtype = np.dtype("bool")
